Task 12.12 done and all data transferred to data.js
