+++ js/data.js
@@ -1,110 +1,84 @@
-import {getRandomPositiveInteger, counterContainer, checkStringLength, shuffle} from './utils.js';
+//server addresses
+const GET_SERVER_ADDRESS = 'https://26.javascript.pages.academy/kekstagram/data';
+const SEND_SERVER_ADDRESS = 'https://26.javascript.pages.academy/kekstagram';
 
-const DESCRIPTIONS = [
-  'Я с друзьями',
-  'Вчера отдыхали, сегодня уже дома',
-  'Я чувствую себя прекрасно! А как у вас дела? Рассказывайте в комментариях!',
-  '~*~*~',
-  'КАК ЖЕ ЭТО КРАСИВО!!1!',
-  'На чилле, на раслабоне'
-];
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-const NAMES = [
-  'Кекс',
-  'Воланд',
-  'Bel0cka',
-  '<3_cool-andrew_2007',
-  'Олег',
-  'Кекс',
-  'Воланд',
-  'Bel0cka',
-  '<3_cool-andrew_2007',
-  'Олег',
-  'Кекс',
-  'Воланд',
-  'Bel0cka',
-  '<3_cool-andrew_2007',
-  'Олег',
-  'Кекс',
-  'Воланд',
-  'Bel0cka',
-  '<3_cool-andrew_2007',
-  'Олег',
-  'Леди Баг'
-];
-
-const PHOTOS_COUNT = 25;
+//photos descriptions constants
+const MAX_HASHTAGS_NUMBER = 5;
 const MAX_COMMENT_LENGTH = 140;
 
-const randomCommentMessagesContainer = function() {
-  const shuffleMESSAGES = shuffle(MESSAGES);
-  let i = 0;
-  return function () {
-    let message = shuffleMESSAGES[i++];
-    if (!checkStringLength(message, MAX_COMMENT_LENGTH)) {
-      message = message.substring(0, MAX_COMMENT_LENGTH);
-    }
-    return message;
-  };
+//filters constants
+const DEBOUNCE_DELAY = 500;
+const NUMBER_OF_PICTURES_FOR_RANDOM_FILTER = 10;
+
+//Pristine object
+const PRISTINE_SETTINGS = {
+  classTo: 'img-upload__field-wrapper',
+  errorClass: 'text--invalid',
+  successClass: 'text--valid',
+  errorTextParent: 'img-upload__field-wrapper',
+  errorTextTag: 'div',
+  errorTextClass: 'text--invalid__error'
 };
 
-const createRandomComments = function() {
-  const commentsCount = getRandomPositiveInteger(0, NAMES.length);
-  const commentsID = counterContainer();
-  const commentMessages = randomCommentMessagesContainer();
+//scale constants
+const ZOOM_STEP = 25;
+const ZOOM_MIN = 25;
+const ZOOM_MAX = 100;
 
-  const commentsArr = [];
-  for (let i = 0; i < commentsCount; i++) {
-    const avatarID = getRandomPositiveInteger(1, NAMES.length);
-    commentsArr[i] = {
-      id: commentsID(),
-      avatar: `img/avatar-${avatarID % 6 + 1}.svg`,
-      message: commentMessages(),
-      name: NAMES[avatarID - 1]
-    };
+//effects list
+const EFFECTS = {
+  chrome: {
+    min: 0,
+    max: 1,
+    step: 0.1,
+    style: 'grayscale',
+    unit: '',
+  },
+  marvin: {
+    min: 0,
+    max: 100,
+    step: 1,
+    style: 'invert',
+    unit: '%',
+  },
+  sepia: {
+    min: 0,
+    max: 1,
+    step: 0.1,
+    style: 'sepia',
+    unit: '',
+  },
+  phobos: {
+    min: 0,
+    max: 3,
+    step: 0.1,
+    style: 'blur',
+    unit: 'px',
+  },
+  heat: {
+    min: 1,
+    max: 3,
+    step: 0.1,
+    style: 'brightness',
+    unit: '',
   }
-
-  return commentsArr;
 };
 
-const getRandomPhotos = function() {
-  const photosID = counterContainer(1);
-  const photosURLID = counterContainer(1);
+//allowed input file types
+const FILE_TYPES = ['gif', 'jpg', 'jpeg', 'png'];
 
-  const photosArr = [];
-  for (let i = 0; i < PHOTOS_COUNT; i++) {
-    photosArr[i] = {
-      id: photosID(),
-      url: `photos/${photosURLID()}.jpg`,
-      description: DESCRIPTIONS[getRandomPositiveInteger(0, DESCRIPTIONS.length - 1)],
-      likes: getRandomPositiveInteger(15, 200),
-      comments: createRandomComments()
-    };
-  }
-
-  return photosArr;
+//export
+export {
+  GET_SERVER_ADDRESS,
+  SEND_SERVER_ADDRESS,
+  MAX_COMMENT_LENGTH,
+  MAX_HASHTAGS_NUMBER,
+  DEBOUNCE_DELAY,
+  NUMBER_OF_PICTURES_FOR_RANDOM_FILTER,
+  PRISTINE_SETTINGS,
+  ZOOM_STEP,
+  ZOOM_MIN,
+  ZOOM_MAX,
+  EFFECTS,
+  FILE_TYPES
 };
-
-export {getRandomPhotos};
